# openapi.yaml
openapi: "3.0.3"
info:
  title: Order Management API
  version: "1.0.0"
  description: Glass Box test fixture — Order Management microservice

servers:
  - url: http://localhost:3000

paths:
  # ─── Orders (source: "both" — matches code endpoints) ───
  /api/orders:
    get:
      summary: List all orders
      operationId: listOrders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Array of orders
    post:
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customerId
                - items
                - shippingAddress
              properties:
                customerId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                shippingAddress:
                  type: string
      responses:
        "201":
          description: Order created
        "400":
          description: Validation error
        "409":
          description: Insufficient inventory

  /api/orders/{id}:
    get:
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
        "404":
          description: Order not found
    put:
      summary: Update order status
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Updated order
        "404":
          description: Order not found
    delete:
      summary: Cancel order
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Order cancelled
        "404":
          description: Order not found
        "409":
          description: Cannot cancel shipped order

  # ─── Products (source: "both") ───
  /api/products:
    get:
      summary: List all products
      operationId: listProducts
      responses:
        "200":
          description: Array of products
    post:
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - sku
                - price
              properties:
                name:
                  type: string
                sku:
                  type: string
                price:
                  type: number
                quantity:
                  type: integer
      responses:
        "201":
          description: Product created
        "409":
          description: Duplicate SKU

  /api/products/{id}:
    get:
      summary: Get product by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
        "404":
          description: Product not found
    put:
      summary: Update product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated product

  # ─── Users (source: "both") ───
  /api/users:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: User created
        "409":
          description: Email already exists

  /api/users/{id}:
    get:
      summary: Get user profile
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
        "404":
          description: User not found
    put:
      summary: Update user profile
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    delete:
      summary: Delete user (admin only)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin only

  # ─── Spec-only endpoints (source: "spec" — NOT in code → spec drift) ───
  /api/legacy/orders:
    get:
      summary: Legacy order listing (deprecated)
      operationId: legacyListOrders
      deprecated: true
      responses:
        "200":
          description: Legacy order array

  /api/reports/sales:
    get:
      summary: Sales report
      operationId: getSalesReport
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales report data

  /api/webhooks/payment:
    post:
      summary: Payment webhook receiver
      operationId: receivePaymentWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed

  # NOTE: /api/health exists in code but NOT in this spec → source: "code" drift
  # NOTE: /api/products/reserve exists in code but NOT in this spec → source: "code" drift